Add files via upload
When learning lists and functions, we use lists where we already know the length of list. If you want to know the max or min of the list, or perform any operation on the list we need to know the length of list, since using range() function needs the length of list. But what if we want the user to enter as many elements they want? This simple script provides a sample program where we can do that. Using 'while True' loop, the list will take as many numbers as possible and prints out the max of the list. Any programmer can use this script and modify according to their needs.